+++ src/view/animationHelper/animationGenerator.js
