+++ src/view/animationHelper/boxModal.js
