+++ src/view/animationHelper/locationCalculator.js
